♻️ refactor(core): extract format-specific submodules
The monolithic __init__.py (~1450 lines) mixed annotation
formatting, type hint resolution, docstring injection,
numpydoc conversion, and Sphinx setup. Extracting focused
modules behind a DocstringFormat ABC makes each concern
independently testable and extensible.

New modules: _annotations.py (leaf, no internal imports),
_resolver.py (type hints, TYPE_CHECKING guards, backfill),
_formats/ (ABC + SphinxFieldListFormat + NumpydocFormat).
The slim __init__.py (~400 lines) orchestrates format
detection and delegates injection. All public symbols are
re-exported for backward compatibility.

Dead code removed during extraction: globals ID set with
no readers, hasattr(__module__) fallback (always True on
3.12+), non-iterable args try/except, Sphinx <9 import
guard (minimum is sphinx>=9.1), unreachable overload and
type_annotation guards. Interpreter-specific branches use
covdefaults markers instead of blanket no-cover pragmas.
