+++ pyproject.toml
@@ -2,7 +2,7 @@
 build-backend = "hatchling.build"
 requires = [
   "hatch-vcs>=0.5",
-  "hatchling>=1.28",
+  "hatchling>=1.29",
 ]
 
 [project]
@@ -60,13 +60,14 @@ test = [
   "coverage>=7.13.4",
   "defusedxml>=0.7.1",
   "diff-cover>=10.2",
+  "numpydoc>=1.8",
   "pytest>=9.0.2",
   "pytest-cov>=7",
   "sphobjinv>=2.3.1.3",
   "typing-extensions>=4.15",
 ]
 type = [
-  "ty>=0.0.17",
+  "ty>=0.0.18",
   { include-group = "docs" },
   { include-group = "test" },
 ]
@@ -74,12 +75,12 @@ docs = [
   "furo>=2025.12.19",
 ]
 lint = [
-  "pre-commit-uv>=4.2",
+  "pre-commit-uv>=4.2.1",
 ]
 pkg-meta = [
   "check-wheel-contents>=0.6.3",
   "twine>=6.2",
-  "uv>=0.10.2",
+  "uv>=0.10.5",
 ]
 
 [tool.hatch]
@@ -119,6 +120,10 @@ lint.per-file-ignores."tests/**/*.py" = [
   "UP007",   # we test for old Union syntax
   "UP045",   # we test for old Optional syntax
 ]
+lint.per-file-ignores."tests/test_resolver.py" = [
+  "ANN001", # type comment fixtures intentionally omit annotations
+  "ARG001", # type comment fixtures have intentionally unused args
+]
 lint.per-file-ignores."tests/roots/test-issue_572/mod_forward_ref.py" = [
   "I002",    # intentionally omits `from __future__ import annotations` to test forward references
   "PLR6301", # method must be instance method to test class forward refs
@@ -146,6 +151,9 @@ max_supported_python = "3.14"
 [tool.pytest]
 ini_options.testpaths = [
   "tests",
+]
+ini_options.filterwarnings = [
+  "ignore:.*rm_rf.*:pytest.PytestWarning",
 ]
 
 [tool.coverage]
