+++ src/sphinx_autodoc_typehints/__init__.py
@@ -2,403 +2,48 @@
 
 from __future__ import annotations
 
-import ast
-import collections.abc
-import enum
-import importlib
 import inspect
-import re
-import sys
-import textwrap
 import types
-from dataclasses import dataclass
-
-if sys.version_info >= (3, 14):
-    import annotationlib
-
-from typing import TYPE_CHECKING, Any, AnyStr, ForwardRef, NewType, TypeVar, Union, get_type_hints
+from typing import TYPE_CHECKING, Any, TypeVar
 
 from docutils import nodes
-from docutils.frontend import get_default_settings
-from docutils.parsers.rst import Directive, directives
-from docutils.utils import new_document
-from sphinx.ext.autodoc.mock import mock
-from sphinx.parsers import RSTParser
-from sphinx.util import logging, rst
-from sphinx.util.docutils import sphinx_domains
-from sphinx.util.inspect import TypeAliasForwardRef, stringify_signature
+from sphinx.util import logging
 from sphinx.util.inspect import signature as sphinx_signature
+from sphinx.util.inspect import stringify_signature
 
-from ._parser import _RstSnippetParser, parse
+# re-exports for backward compatibility
+from ._annotations import (
+    MyTypeAliasForwardRef,
+    add_type_css_class,
+    format_annotation,
+    get_annotation_args,
+    get_annotation_class_name,
+    get_annotation_module,
+    unescape,
+)
+from ._formats import detect_format
+from ._formats._numpydoc import _convert_numpydoc_to_sphinx_fields  # noqa: F401
+from ._formats._sphinx import _has_yields_section, _is_generator_type
+from ._parser import parse
+from ._resolver import (
+    _collect_documented_type_aliases,
+    backfill_type_hints,
+    get_all_type_hints,
+    normalize_source_lines,
+)
 from .patches import _OVERLOADS_CACHE, install_patches
 from .version import __version__
 
 if TYPE_CHECKING:
-    import optparse
-    from ast import FunctionDef, Module, stmt
     from collections.abc import Callable
 
-    from docutils.frontend import Values
     from docutils.nodes import Node
     from docutils.parsers.rst import states
     from sphinx.application import Sphinx
-    from sphinx.config import Config
     from sphinx.environment import BuildEnvironment
     from sphinx.ext.autodoc import Options
 
 _LOGGER = logging.getLogger(__name__)
-
-_BUILTIN_DIRECTIVES = frozenset(directives._directive_registry)  # noqa: SLF001
-_PYDATA_ANNOTS_TYPING = {
-    "Any",
-    "AnyStr",
-    "Callable",
-    "ClassVar",
-    "Literal",
-    "NoReturn",
-    "Optional",
-    "Tuple",
-    *({"Union"} if sys.version_info < (3, 14) else set()),
-}
-_PYDATA_ANNOTS_TYPES = {
-    *("AsyncGeneratorType", "BuiltinFunctionType", "BuiltinMethodType"),
-    *("CellType", "ClassMethodDescriptorType", "CoroutineType"),
-    "EllipsisType",
-    *("FrameType", "FunctionType"),
-    *("GeneratorType", "GetSetDescriptorType"),
-    "LambdaType",
-    *("MemberDescriptorType", "MethodDescriptorType", "MethodType", "MethodWrapperType"),
-    # NoneType is special, but included here for completeness' sake
-    *("NoneType", "NotImplementedType"),
-    "WrapperDescriptorType",
-}
-_PYDATA_ANNOTATIONS = {
-    *(("typing", n) for n in _PYDATA_ANNOTS_TYPING),
-    *(("types", n) for n in _PYDATA_ANNOTS_TYPES),
-}
-
-# types has a bunch of things like ModuleType where ModuleType.__module__ is
-# "builtins" and ModuleType.__name__ is "module", so we have to check for this.
-_TYPES_DICT = {getattr(types, name): name for name in types.__all__}
-# Prefer FunctionType to LambdaType (they are synonymous)
-_TYPES_DICT[types.FunctionType] = "FunctionType"
-
-
-class MyTypeAliasForwardRef(TypeAliasForwardRef):
-    def __or__(self, value: Any) -> Any:  # ty: ignore[invalid-method-override]
-        return Union[self, value]  # noqa: UP007
-
-
-def _get_types_type(obj: Any) -> str | None:
-    try:
-        return _TYPES_DICT.get(obj)
-    except Exception:  # noqa: BLE001
-        # e.g. exception: unhashable type
-        return None
-
-
-def get_annotation_module(annotation: Any) -> str:
-    """
-    Get module for an annotation.
-
-    :param annotation:
-    :return:
-    """
-    if annotation is None:
-        return "builtins"
-    if _get_types_type(annotation) is not None:
-        return "types"
-    is_new_type = isinstance(annotation, NewType)
-    if (
-        is_new_type
-        or isinstance(annotation, TypeVar)
-        or type(annotation).__name__ in {"ParamSpec", "ParamSpecArgs", "ParamSpecKwargs"}
-    ):
-        return "typing"
-    if hasattr(annotation, "__module__"):
-        return annotation.__module__
-    if hasattr(annotation, "__origin__"):
-        return annotation.__origin__.__module__
-    msg = f"Cannot determine the module of {annotation}"
-    raise ValueError(msg)
-
-
-def _is_newtype(annotation: Any) -> bool:
-    return isinstance(annotation, NewType)
-
-
-def get_annotation_class_name(annotation: Any, module: str) -> str:  # noqa: C901, PLR0911
-    """
-    Get class name for annotation.
-
-    :param annotation:
-    :param module:
-    :return:
-    """
-    # Special cases
-    if annotation is None:
-        return "None"
-    if annotation is AnyStr:
-        return "AnyStr"
-    val = _get_types_type(annotation)
-    if val is not None:
-        return val
-    if _is_newtype(annotation):
-        return "NewType"
-
-    if getattr(annotation, "__qualname__", None):
-        return annotation.__qualname__
-    if getattr(annotation, "_name", None):  # Required for generic aliases on Python 3.7+
-        return annotation._name  # noqa: SLF001
-    if module in {"typing", "typing_extensions"} and isinstance(getattr(annotation, "name", None), str):
-        # Required for at least Pattern and Match
-        return annotation.name
-
-    origin = getattr(annotation, "__origin__", None)
-    if origin:
-        if getattr(origin, "__qualname__", None):  # Required for Protocol subclasses
-            return origin.__qualname__
-        if getattr(origin, "_name", None):  # Required for Union on Python 3.7+
-            return origin._name  # noqa: SLF001
-
-    annotation_cls = annotation if inspect.isclass(annotation) else type(annotation)
-    return annotation_cls.__qualname__.lstrip("_")
-
-
-def get_annotation_args(annotation: Any, module: str, class_name: str) -> tuple[Any, ...]:  # noqa: PLR0911
-    """
-    Get annotation arguments.
-
-    :param annotation:
-    :param module:
-    :param class_name:
-    :return:
-    """
-    try:
-        original = getattr(sys.modules[module], class_name)
-    except (KeyError, AttributeError):
-        pass
-    else:
-        if annotation is original:
-            return ()  # This is the original, not parametrized type
-
-    # Special cases
-    if class_name in {"Pattern", "Match"} and hasattr(annotation, "type_var"):  # Python < 3.7
-        return (annotation.type_var,)
-    if class_name == "ClassVar" and hasattr(annotation, "__type__"):  # ClassVar on Python < 3.7
-        return (annotation.__type__,)
-    if class_name == "TypeVar" and hasattr(annotation, "__constraints__"):
-        return annotation.__constraints__
-    if class_name == "NewType" and hasattr(annotation, "__supertype__"):
-        return (annotation.__supertype__,)
-    if class_name == "Literal" and hasattr(annotation, "__values__"):
-        return annotation.__values__
-    if class_name == "Generic":
-        return annotation.__parameters__
-    result = getattr(annotation, "__args__", ())
-    # 3.10 and earlier Tuple[()] returns ((), ) instead of () the tuple does
-    return () if len(result) == 1 and result[0] == () else result  # type: ignore[misc]
-
-
-def format_internal_tuple(t: tuple[Any, ...], config: Config, *, short_literals: bool = False) -> str:
-    # An annotation can be a tuple, e.g., for numpy.typing:
-    # In this case, format_annotation receives:
-    # This solution should hopefully be general for *any* type that allows tuples in annotations
-    fmt = [format_annotation(a, config, short_literals=short_literals) for a in t]
-    if len(fmt) == 0:
-        return "()"
-    if len(fmt) == 1:
-        return f"({fmt[0]}, )"
-    return f"({', '.join(fmt)})"
-
-
-def fixup_module_name(config: Config, module: str) -> str:
-    if getattr(config, "typehints_fixup_module_name", None):
-        module = config.typehints_fixup_module_name(module)
-
-    if module == "typing_extensions":
-        module = "typing"
-
-    if module == "_io":
-        module = "io"
-    return module
-
-
-def _format_literal_arg(arg: Any, config: Config) -> str:
-    if isinstance(arg, enum.Enum):
-        enum_cls = type(arg)
-        module = fixup_module_name(config, enum_cls.__module__)
-        fully_qualified = getattr(config, "typehints_fully_qualified", False)
-        qualified = f"{module}.{enum_cls.__qualname__}.{arg.name}" if module else f"{enum_cls.__qualname__}.{arg.name}"
-        prefix = "" if fully_qualified or not module else "~"
-        return f":py:attr:`{prefix}{qualified}`"
-    return f"``{arg!r}``"
-
-
-def format_annotation(annotation: Any, config: Config, *, short_literals: bool = False) -> str:  # noqa: C901, PLR0911, PLR0912, PLR0915, PLR0914
-    """
-    Format the annotation.
-
-    :param annotation:
-    :param config:
-    :param short_literals: Render :py:class:`Literals` in PEP 604 style (``|``).
-    :return:
-    """
-    typehints_formatter: Callable[..., str] | None = getattr(config, "typehints_formatter", None)
-    if typehints_formatter is not None:
-        formatted = typehints_formatter(annotation, config)
-        if formatted is not None:
-            return formatted
-
-    # Special cases
-    if isinstance(annotation, ForwardRef):
-        return annotation.__forward_arg__
-    if annotation is None or annotation is type(None):
-        return ":py:obj:`None`"
-    if annotation is Ellipsis:
-        return ":py:data:`...<Ellipsis>`"
-
-    if isinstance(annotation, tuple):
-        return format_internal_tuple(annotation, config)
-
-    if isinstance(annotation, TypeAliasForwardRef):
-        if (env := getattr(config, "_typehints_env", None)) is not None:
-            py_domain = env.get_domain("py")
-            module_prefix = getattr(config, "_typehints_module_prefix", "")
-            for candidate in (f"{module_prefix}.{annotation.name}", annotation.name):
-                if candidate in py_domain.objects and py_domain.objects[candidate].objtype == "type":
-                    fully_qualified: bool = getattr(config, "typehints_fully_qualified", False)
-                    prefix = "" if fully_qualified else "~"
-                    return f":py:type:`{prefix}{candidate}`"
-        return annotation.name
-
-    try:
-        module = get_annotation_module(annotation)
-        class_name = get_annotation_class_name(annotation, module)
-        args = get_annotation_args(annotation, module, class_name)
-    except ValueError:
-        return str(annotation).strip("'")
-
-    module = fixup_module_name(config, module)
-    full_name = f"{module}.{class_name}" if module != "builtins" else class_name
-    fully_qualified: bool = getattr(config, "typehints_fully_qualified", False)
-    prefix = "" if fully_qualified or full_name == class_name else "~"
-    role = "data" if (module, class_name) in _PYDATA_ANNOTATIONS else "class"
-    args_format = "\\[{}]"
-    formatted_args: str | None = ""
-
-    always_use_bars_union: bool = getattr(config, "always_use_bars_union", True)
-    is_bars_union = (
-        (sys.version_info >= (3, 14) and full_name == "typing.Union")
-        or full_name == "types.UnionType"
-        or (always_use_bars_union and type(annotation).__qualname__ == "_UnionGenericAlias")
-    )
-    if is_bars_union:
-        full_name = ""
-
-    # Some types require special handling
-    if full_name == "typing.NewType":
-        newtype_module = fixup_module_name(config, getattr(annotation, "__module__", ""))
-        newtype_name = annotation.__name__
-        newtype_qualified = f"{newtype_module}.{newtype_name}" if newtype_module else newtype_name
-        newtype_prefix = "" if fully_qualified or not newtype_module else "~"
-        supertype = format_annotation(annotation.__supertype__, config, short_literals=short_literals)
-        return f":py:class:`{newtype_prefix}{newtype_qualified}` ({supertype})"
-    if full_name == "typing.Annotated":
-        # By default we don't show metadata in Annotated
-        return format_annotation(annotation.__origin__, config, short_literals=short_literals)
-    if full_name in {"typing.TypeVar", "typing.ParamSpec"}:
-        params = {k: getattr(annotation, f"__{k}__") for k in ("bound", "covariant", "contravariant")}
-        params = {k: v for k, v in params.items() if v}
-        if "bound" in params:
-            params["bound"] = f" {format_annotation(params['bound'], config, short_literals=short_literals)}"
-        args_format = f"\\(``{annotation.__name__}``{', {}' if args else ''}"
-        if params:
-            args_format += "".join(f", {k}={v}" for k, v in params.items())
-        args_format += ")"
-        formatted_args = None if args else args_format
-    elif full_name == "typing.Optional":
-        args = tuple(x for x in args if x is not type(None))
-    elif full_name in {"typing.Union", "types.UnionType"} and type(None) in args:
-        if len(args) == 2:  # noqa: PLR2004
-            full_name = "typing.Optional"
-            role = "data"
-            args = tuple(x for x in args if x is not type(None))
-        else:
-            simplify_optional_unions: bool = getattr(config, "simplify_optional_unions", True)
-            if not simplify_optional_unions:
-                full_name = "typing.Optional"
-                role = "data"
-                args_format = f"\\[:py:data:`{prefix}typing.Union`\\[{{}}]]"
-                args = tuple(x for x in args if x is not type(None))
-    elif full_name in {"typing.Callable", "collections.abc.Callable"} and args and args[0] is not ...:
-        fmt = [format_annotation(arg, config, short_literals=short_literals) for arg in args]
-        formatted_args = f"\\[\\[{', '.join(fmt[:-1])}], {fmt[-1]}]"
-    elif full_name == "typing.Literal":
-        literal_parts = [_format_literal_arg(arg, config) for arg in args]
-        if short_literals:
-            return f"\\{' | '.join(literal_parts)}"
-        formatted_args = f"\\[{', '.join(literal_parts)}]"
-    elif is_bars_union:
-        if not args:
-            return f":py:{'class' if sys.version_info >= (3, 14) else 'data'}:`{prefix}typing.Union`"
-        return " | ".join([format_annotation(arg, config, short_literals=short_literals) for arg in args])
-
-    if args and not formatted_args:
-        try:
-            iter(args)
-        except TypeError:
-            fmt = [format_annotation(args, config, short_literals=short_literals)]
-        else:
-            fmt = [format_annotation(arg, config, short_literals=short_literals) for arg in args]
-        formatted_args = args_format.format(", ".join(fmt))
-
-    escape = "\\ " if formatted_args else ""
-    return f":py:{role}:`{prefix}{full_name}`{escape}{formatted_args}"
-
-
-# reference: https://github.com/pytorch/pytorch/pull/46548/files
-def normalize_source_lines(source_lines: str) -> str:
-    """
-    Normalize the source lines.
-
-    It finds the indentation level of the function definition (`def`), then it indents all lines in the function body to
-    a point at or greater than that level. This allows for comments and continued string literals that are at a lower
-    indentation than the rest of the code.
-
-    :param source_lines: source code
-    :return: source lines that have been correctly aligned
-    """
-    lines = source_lines.split("\n")
-
-    def remove_prefix(text: str, prefix: str) -> str:
-        return text[text.startswith(prefix) and len(prefix) :]
-
-    # Find the line and line number containing the function definition
-    for pos, line in enumerate(lines):
-        if line.lstrip().startswith("def "):
-            idx = pos
-            whitespace_separator = "def"
-            break
-        if line.lstrip().startswith("async def"):
-            idx = pos
-            whitespace_separator = "async def"
-            break
-
-    else:
-        return "\n".join(lines)
-    fn_def = lines[idx]
-
-    # Get a string representing the amount of leading whitespace
-    whitespace = fn_def.split(whitespace_separator)[0]
-
-    # Add this leading whitespace to all lines before and after the `def`
-    aligned_prefix = [whitespace + remove_prefix(s, whitespace) for s in lines[:idx]]
-    aligned_suffix = [whitespace + remove_prefix(s, whitespace) for s in lines[idx + 1 :]]
-
-    # Put it together again
-    aligned_prefix.append(fn_def)
-    return "\n".join(aligned_prefix + aligned_suffix)
 
 
 def process_signature(  # noqa: C901, PLR0911, PLR0912, PLR0913, PLR0917
@@ -410,24 +55,13 @@ def process_signature(  # noqa: C901, PLR0911, PLR0912, PLR0913, PLR0917
     signature: str,  # noqa: ARG001
     return_annotation: str,  # noqa: ARG001
 ) -> tuple[str, None] | None:
-    """
-    Process the signature.
-
-    :param app:
-    :param what:
-    :param name:
-    :param obj:
-    :param options:
-    :param signature:
-    :param return_annotation:
-    :return:
-    """
+    """Process the signature."""
     if not callable(obj):
         return None
 
     original_obj = obj
     obj = getattr(obj, "__init__", getattr(obj, "__new__", None)) if inspect.isclass(obj) else obj
-    if not getattr(obj, "__annotations__", None):  # when has no annotation we cannot autodoc typehints so bail
+    if not getattr(obj, "__annotations__", None):
         return None
 
     try:
@@ -456,13 +90,11 @@ def process_signature(  # noqa: C901, PLR0911, PLR0912, PLR0913, PLR0917
     else:
         parameters = [param.replace(annotation=inspect.Parameter.empty) for param in sph_signature.parameters.values()]
 
-    # if we have parameters we may need to delete first argument that's not documented, e.g. self
     start = 0
     if parameters:
         if inspect.isclass(original_obj) or (what == "method" and name.endswith(".__init__")):
             start = 1
         elif what == "method":
-            # bail if it is a local method as we cannot determine if first argument needs to be deleted or not
             if "<locals>" in obj.__qualname__ and not _is_dataclass(name, what, obj.__qualname__):
                 _LOGGER.warning(
                     'Cannot handle as a local function: "%s" (use @functools.wraps)',
@@ -479,7 +111,6 @@ def process_signature(  # noqa: C901, PLR0911, PLR0912, PLR0913, PLR0917
                     return None
             method_name = obj.__name__
             if method_name.startswith("__") and not method_name.endswith("__"):
-                # when method starts with double underscore Python applies mangling -> prepend the class name
                 method_name = f"_{obj.__qualname__.split('.')[-2]}{method_name}"
             method_object = outer.__dict__.get(method_name, obj) if outer else obj
             if not isinstance(method_object, classmethod | staticmethod):
@@ -499,283 +130,7 @@ def process_signature(  # noqa: C901, PLR0911, PLR0912, PLR0913, PLR0917
 
 
 def _is_dataclass(name: str, what: str, qualname: str) -> bool:
-    # generated dataclass __init__() and class need extra checks, as the function operates on the generated class
-    # and methods (not an instantiated dataclass object) it cannot be replaced by a call to
-    # `dataclasses.is_dataclass()` => check manually for either generated __init__ or generated class
     return (what == "method" and name.endswith(".__init__")) or (what == "class" and qualname.endswith(".__init__"))
-
-
-def _future_annotations_imported(obj: Any) -> bool:
-    annotations_ = getattr(inspect.getmodule(obj), "annotations", None)
-    if annotations_ is None:
-        return False
-
-    # Make sure that annotations is imported from __future__ - defined in cpython/Lib/__future__.py
-    # annotations become strings at runtime
-    return bool(annotations_.compiler_flag == 0x1000000)  # pragma: no cover # noqa: PLR2004
-
-
-def get_all_type_hints(
-    autodoc_mock_imports: list[str], obj: Any, name: str, localns: dict[Any, MyTypeAliasForwardRef]
-) -> dict[str, Any]:
-    result = _get_type_hint(autodoc_mock_imports, name, obj, localns)
-    if not result:
-        result = backfill_type_hints(obj, name)
-        try:
-            obj.__annotations__ = result
-        except (AttributeError, TypeError):
-            pass
-        else:
-            result = _get_type_hint(autodoc_mock_imports, name, obj, localns)
-    return result
-
-
-_TYPE_GUARD_IMPORT_RE = re.compile(r"\nif (typing.)?TYPE_CHECKING:[^\n]*([\s\S]*?)(?=\n\S)")
-_TYPE_GUARD_IMPORTS_RESOLVED = set()
-_TYPE_GUARD_IMPORTS_RESOLVED_GLOBALS_ID = set()
-
-
-def _should_skip_guarded_import_resolution(obj: Any) -> bool:
-    if isinstance(obj, types.ModuleType):
-        return False  # Don't skip modules
-
-    if not hasattr(obj, "__globals__"):
-        return True  # Skip objects without __globals__
-
-    if hasattr(obj, "__module__"):
-        return obj.__module__ in _TYPE_GUARD_IMPORTS_RESOLVED or obj.__module__ in sys.builtin_module_names
-
-    return id(obj.__globals__) in _TYPE_GUARD_IMPORTS_RESOLVED_GLOBALS_ID
-
-
-def _execute_guarded_code(autodoc_mock_imports: list[str], obj: Any, module_code: str) -> None:
-    for _, part in _TYPE_GUARD_IMPORT_RE.findall(module_code):
-        guarded_code = textwrap.dedent(part)
-        try:
-            _run_guarded_import(autodoc_mock_imports, obj, guarded_code)
-        except Exception as exc:  # noqa: BLE001
-            _LOGGER.warning(
-                "Failed guarded type import with %r", exc, type="sphinx_autodoc_typehints", subtype="guarded_import"
-            )
-
-
-def _run_guarded_import(autodoc_mock_imports: list[str], obj: Any, guarded_code: str) -> None:
-    ns = getattr(obj, "__globals__", obj.__dict__)
-    try:
-        with mock(autodoc_mock_imports):
-            exec(guarded_code, ns)  # noqa: S102
-    except ImportError as exc:
-        if not exc.name:
-            return
-        _resolve_type_guarded_imports(autodoc_mock_imports, importlib.import_module(exc.name))
-        try:
-            with mock(autodoc_mock_imports):
-                exec(guarded_code, ns)  # noqa: S102
-        except ImportError:
-            pass  # third-party internal import that doesn't exist at runtime
-
-
-def _resolve_type_guarded_imports(autodoc_mock_imports: list[str], obj: Any) -> None:
-    if _should_skip_guarded_import_resolution(obj):
-        return
-
-    if hasattr(obj, "__globals__"):
-        _TYPE_GUARD_IMPORTS_RESOLVED_GLOBALS_ID.add(id(obj.__globals__))
-
-    module = inspect.getmodule(obj)
-
-    if module:
-        try:
-            module_code = inspect.getsource(module)
-        except (TypeError, OSError):
-            ...  # no source code => no type guards
-        else:
-            _TYPE_GUARD_IMPORTS_RESOLVED.add(module.__name__)
-            _execute_guarded_code(autodoc_mock_imports, obj, module_code)
-
-
-def _add_type_params_to_localns(
-    obj: Any, localns: dict[Any, MyTypeAliasForwardRef]
-) -> dict[Any, MyTypeAliasForwardRef]:
-    if type_params := getattr(obj, "__type_params__", None):
-        localns = {**localns, **{p.__name__: p for p in type_params}}
-    qualname = getattr(obj, "__qualname__", "") or ""
-    parts = qualname.rsplit(".", 1)
-    if len(parts) > 1:
-        parent_name = parts[0]
-        ns = getattr(obj, "__globals__", None)
-        if ns is None:
-            module = inspect.getmodule(obj)
-            ns = vars(module) if module else None
-        if ns and (parent := ns.get(parent_name)) and (parent_params := getattr(parent, "__type_params__", None)):
-            localns = {**localns, **{p.__name__: p for p in parent_params}}
-    return localns
-
-
-def _get_type_hint(
-    autodoc_mock_imports: list[str], name: str, obj: Any, localns: dict[Any, MyTypeAliasForwardRef]
-) -> dict[str, Any]:
-    _resolve_type_guarded_imports(autodoc_mock_imports, obj)
-    localns = _add_type_params_to_localns(obj, localns)
-    try:
-        result = get_type_hints(obj, None, localns, include_extras=True)
-    except (AttributeError, TypeError, RecursionError) as exc:
-        # TypeError - slot wrapper, PEP-563 when part of new syntax not supported
-        # RecursionError - some recursive type definitions https://github.com/python/typing/issues/574
-        if isinstance(exc, TypeError) and _future_annotations_imported(obj) and "unsupported operand type" in str(exc):
-            result = obj.__annotations__
-        else:
-            result = {}
-    except NameError as exc:
-        _LOGGER.warning(
-            'Cannot resolve forward reference in type annotations of "%s": %s',
-            name,
-            exc,
-            type="sphinx_autodoc_typehints",
-            subtype="forward_reference",
-        )
-        if sys.version_info >= (3, 14):
-            result = annotationlib.get_annotations(obj, format=annotationlib.Format.FORWARDREF)
-        else:
-            result = obj.__annotations__
-    return result
-
-
-def backfill_type_hints(obj: Any, name: str) -> dict[str, Any]:  # noqa: C901, PLR0911
-    """
-    Backfill type hints.
-
-    :param obj: the object
-    :param name: the name
-    :return: backfilled value
-    """
-    parse_kwargs = {"type_comments": True}
-
-    def _one_child(module: Module) -> stmt | None:
-        children = module.body  # use the body to ignore type comments
-        if len(children) != 1:
-            _LOGGER.warning(
-                'Did not get exactly one node from AST for "%s", got %s',
-                name,
-                len(children),
-                type="sphinx_autodoc_typehints",
-                subtype="multiple_ast_nodes",
-            )
-            return None
-        return children[0]
-
-    try:
-        code = textwrap.dedent(normalize_source_lines(inspect.getsource(obj)))
-        obj_ast = ast.parse(code, **parse_kwargs)  # type: ignore[call-overload]  # dynamic kwargs
-    except (OSError, TypeError, SyntaxError):
-        return {}
-
-    obj_ast = _one_child(obj_ast)
-    if obj_ast is None:
-        return {}
-
-    try:
-        type_comment = obj_ast.type_comment  # type: ignore[attr-defined]
-    except AttributeError:
-        return {}
-
-    if not type_comment:
-        return {}
-
-    try:
-        comment_args_str, comment_returns = type_comment.split(" -> ")
-    except ValueError:
-        _LOGGER.warning(
-            'Unparseable type hint comment for "%s": Expected to contain ` -> `',
-            name,
-            type="sphinx_autodoc_typehints",
-            subtype="comment",
-        )
-        return {}
-
-    rv = {}
-    if comment_returns:
-        rv["return"] = comment_returns
-
-    args = load_args(obj_ast)  # type: ignore[arg-type]
-    comment_args = split_type_comment_args(comment_args_str)
-    is_inline = len(comment_args) == 1 and comment_args[0] == "..."
-    if not is_inline:
-        if args and args[0].arg in {"self", "cls"} and len(comment_args) != len(args):
-            comment_args.insert(0, None)  # self/cls may be omitted in type comments, insert blank
-
-        if len(args) != len(comment_args):
-            _LOGGER.warning(
-                'Not enough type comments found on "%s"', name, type="sphinx_autodoc_typehints", subtype="comment"
-            )
-            return rv
-
-    for at, arg in enumerate(args):
-        arg_key = getattr(arg, "arg", None)
-        if arg_key is None:
-            continue
-
-        value = getattr(arg, "type_comment", None) if is_inline else comment_args[at]
-
-        if value is not None:
-            rv[arg_key] = value
-
-    return rv
-
-
-def load_args(obj_ast: FunctionDef) -> list[Any]:
-    func_args = obj_ast.args
-    args = []
-    pos_only = getattr(func_args, "posonlyargs", None)
-    if pos_only:
-        args.extend(pos_only)
-
-    args.extend(func_args.args)
-    if func_args.vararg:
-        args.append(func_args.vararg)
-
-    args.extend(func_args.kwonlyargs)
-    if func_args.kwarg:
-        args.append(func_args.kwarg)
-
-    return args
-
-
-def split_type_comment_args(comment: str) -> list[str | None]:
-    def add(val: str) -> None:
-        result.append(val.strip().lstrip("*"))  # remove spaces, and var/kw arg marker
-
-    comment = comment.strip().lstrip("(").rstrip(")")
-    result: list[str | None] = []
-    if not comment:
-        return result
-
-    brackets, start_arg_at, at = 0, 0, 0
-    for at, char in enumerate(comment):
-        if char in {"[", "("}:
-            brackets += 1
-        elif char in {"]", ")"}:
-            brackets -= 1
-        elif char == "," and brackets == 0:
-            add(comment[start_arg_at:at])
-            start_arg_at = at + 1
-
-    add(comment[start_arg_at : at + 1])
-    return result
-
-
-def format_default(app: Sphinx, default: Any, is_annotated: bool) -> str | None:  # noqa: FBT001
-    if default is inspect.Parameter.empty:
-        return None
-    formatted = repr(default).replace("\\", "\\\\")
-
-    if is_annotated:
-        if app.config.typehints_defaults.startswith("braces"):
-            return f" (default: ``{formatted}``)"
-        return f", default: ``{formatted}``"
-    if app.config.typehints_defaults == "braces-after":
-        return f" (default: ``{formatted}``)"
-    return f"default: ``{formatted}``"
 
 
 def process_docstring(  # noqa: PLR0913, PLR0917
@@ -786,17 +141,7 @@ def process_docstring(  # noqa: PLR0913, PLR0917
     options: Options | None,  # noqa: ARG001
     lines: list[str],
 ) -> None:
-    """
-    Process the docstring for an entry.
-
-    :param app: the Sphinx app
-    :param what: the target
-    :param name: the name
-    :param obj: the object
-    :param options: the options
-    :param lines: the lines
-    :return:
-    """
+    """Process the docstring for an entry."""
     original_obj = obj
     obj = obj.fget if isinstance(obj, property) else obj
     if not callable(obj):
@@ -834,56 +179,7 @@ def process_docstring(  # noqa: PLR0913, PLR0917
         delattr(app.config, "_typehints_module_prefix")
 
 
-def _collect_documented_type_aliases(
-    obj: Any, module_prefix: str, env: BuildEnvironment
-) -> tuple[dict[str, MyTypeAliasForwardRef], dict[int, MyTypeAliasForwardRef]]:
-    """
-    Find annotation names that reference documented ``.. py:type::`` entries.
-
-    Scans the function's raw ``__annotations__`` to find type names before ``get_type_hints()`` expands them.
-    For deferred annotations (strings), checks the name directly and returns them in ``localns`` for
-    ``get_type_hints()`` to resolve. For already-resolved annotations (no ``from __future__ import annotations``),
-    scans module globals to find variable names pointing to the same type object, and returns them keyed by
-    ``id()`` for post-processing after ``get_type_hints()``.
-
-    :param obj: The callable whose annotations to scan
-    :param module_prefix: The module name to prefix type names with (e.g. "mod")
-    :param env: The Sphinx build environment
-    :return: ``(deferred, eager)`` where deferred maps name to ref for localns, eager maps id to ref for post-processing
-    """
-    raw_annotations = getattr(obj, "__annotations__", {})
-    if not raw_annotations:
-        return {}, {}
-
-    py_objects = env.get_domain("py").objects  # ty: ignore[unresolved-attribute]
-    deferred: dict[str, MyTypeAliasForwardRef] = {}
-    eager: dict[int, MyTypeAliasForwardRef] = {}
-    obj_globals = getattr(obj, "__globals__", {})
-
-    for annotation in raw_annotations.values():
-        if isinstance(annotation, str):
-            if _is_documented_type(annotation, module_prefix, py_objects):
-                deferred[annotation] = MyTypeAliasForwardRef(annotation)
-        else:
-            for var_name, var_value in obj_globals.items():
-                if (
-                    var_value is annotation
-                    and not var_name.startswith("_")
-                    and _is_documented_type(var_name, module_prefix, py_objects)
-                ):
-                    eager[id(annotation)] = MyTypeAliasForwardRef(var_name)
-
-    return deferred, eager
-
-
-def _is_documented_type(name: str, module_prefix: str, py_objects: dict[str, Any]) -> bool:
-    return any(
-        candidate in py_objects and py_objects[candidate].objtype == "type"
-        for candidate in (f"{module_prefix}.{name}", name)
-    )
-
-
-def _inject_overload_signatures(  # noqa: C901
+def _inject_overload_signatures(
     app: Sphinx,
     what: str,
     name: str,  # noqa: ARG001
@@ -928,54 +224,24 @@ def _inject_overload_signatures(  # noqa: C901
         sig_line = f"   * {', '.join(params)}{return_annotation}"
         overload_lines.append(sig_line)
 
-    if len(overload_lines) > 1:
-        overload_lines.append("")
-        for line in reversed(overload_lines):
-            lines.insert(0, line)
-        return True
-
-    return False
+    overload_lines.append("")
+    for line in reversed(overload_lines):
+        lines.insert(0, line)
+    return True
 
 
-def _get_sphinx_line_keyword_and_argument(line: str) -> tuple[str, str | None] | None:
-    """
-    Extract a keyword, and its optional argument out of a sphinx field option line.
-
-    For example
-    >>> _get_sphinx_line_keyword_and_argument(":param parameter:")
-    ("param", "parameter")
-    >>> _get_sphinx_line_keyword_and_argument(":return:")
-    ("return", None)
-    >>> _get_sphinx_line_keyword_and_argument("some invalid line")
-    None
-    """
-    param_line_without_description = line.split(":", maxsplit=2)
-    if len(param_line_without_description) != 3:  # noqa: PLR2004
+def format_default(app: Sphinx, default: Any, is_annotated: bool) -> str | None:  # noqa: FBT001
+    if default is inspect.Parameter.empty:
         return None
+    formatted = repr(default).replace("\\", "\\\\")
 
-    split_directive_and_name = param_line_without_description[1].split(maxsplit=1)
-    if len(split_directive_and_name) != 2:  # noqa: PLR2004
-        if not len(split_directive_and_name):
-            return None
-        return split_directive_and_name[0], None
-
-    return tuple(split_directive_and_name)  # type: ignore[return-value]
-
-
-def _line_is_param_line_for_arg(line: str, arg_name: str) -> bool:
-    """Return True if `line` is a valid parameter line for `arg_name`, false otherwise."""
-    keyword_and_name = _get_sphinx_line_keyword_and_argument(line)
-    if keyword_and_name is None:
-        return False
-
-    keyword, doc_name = keyword_and_name
-    if doc_name is None:
-        return False
-
-    if keyword not in {"param", "parameter", "arg", "argument"}:
-        return False
-
-    return any(doc_name == prefix + arg_name for prefix in ("", "\\*", "\\**", "\\*\\*"))
+    if is_annotated:
+        if app.config.typehints_defaults.startswith("braces"):
+            return f" (default: ``{formatted}``)"
+        return f", default: ``{formatted}``"
+    if app.config.typehints_defaults == "braces-after":
+        return f" (default: ``{formatted}``)"
+    return f"default: ``{formatted}``"
 
 
 def _inject_types_to_docstring(  # noqa: PLR0913, PLR0917
@@ -988,10 +254,11 @@ def _inject_types_to_docstring(  # noqa: PLR0913, PLR0917
     lines: list[str],
     has_overloads: bool = False,  # noqa: FBT001, FBT002
 ) -> None:
+    fmt = detect_format(lines)
     if signature is not None:
-        _inject_signature(type_hints, signature, app, lines)
+        _inject_signature(type_hints, signature, app, lines, fmt)
     if "return" in type_hints and not has_overloads:
-        _inject_rtype(type_hints, original_obj, app, what, name, lines)
+        _inject_rtype(type_hints, original_obj, app, what, name, lines, fmt)
 
 
 def _inject_signature(
@@ -999,43 +266,39 @@ def _inject_signature(
     signature: inspect.Signature,
     app: Sphinx,
     lines: list[str],
+    fmt: Any,
 ) -> None:
     for arg_name in signature.parameters:
-        _inject_arg_signature(type_hints, signature, app, lines, arg_name)
+        _inject_arg_signature(type_hints, signature, app, lines, arg_name, fmt)
 
 
-def _inject_arg_signature(
+def _inject_arg_signature(  # noqa: PLR0913, PLR0917
     type_hints: dict[str, Any],
     signature: inspect.Signature,
     app: Sphinx,
     lines: list[str],
     arg_name: str,
+    fmt: Any,
 ) -> None:
     annotation = type_hints.get(arg_name)
-
     default = signature.parameters[arg_name].default
 
     if arg_name.endswith("_"):
         arg_name = f"{arg_name[:-1]}\\_"
 
-    insert_index = None
-    for at, line in enumerate(lines):
-        if _line_is_param_line_for_arg(line, arg_name):
-            # Get the arg_name from the doc to match up for type in case it has a star prefix.
-            # Line is in the correct format so this is guaranteed to return tuple[str, str].
-            _, arg_name = _get_sphinx_line_keyword_and_argument(line)  # type: ignore[assignment, misc]
-            insert_index = at
-            break
+    insert_index = fmt.find_param(lines, arg_name)
+
+    if insert_index is not None and hasattr(fmt, "get_arg_name_from_line"):
+        arg_name = fmt.get_arg_name_from_line(lines[insert_index]) or arg_name
 
     if annotation is not None and insert_index is None and app.config.always_document_param_types:
-        lines.append(f":param {arg_name}:")
-        insert_index = len(lines) - 1
+        insert_index = fmt.add_undocumented_param(lines, arg_name)
 
     if insert_index is not None:
         has_preexisting_annotation = False
 
         if annotation is None:
-            type_annotation, has_preexisting_annotation = _find_preexisting_type_annotation(lines, arg_name)
+            type_annotation, has_preexisting_annotation = fmt.find_preexisting_type(lines, arg_name)
         else:
             short_literals = app.config.python_display_short_literal_types
             formatted_annotation = add_type_css_class(
@@ -1046,163 +309,26 @@ def _inject_arg_signature(
         if app.config.typehints_defaults:
             formatted_default = format_default(app, default, annotation is not None or has_preexisting_annotation)
             if formatted_default:
-                type_annotation = _append_default(app, lines, insert_index, type_annotation, formatted_default)
+                after = app.config.typehints_defaults.endswith("after")
+                type_annotation = fmt.append_default(
+                    lines, insert_index, type_annotation, formatted_default, after=after
+                )
 
         lines.insert(insert_index, type_annotation)
 
 
-def _find_preexisting_type_annotation(lines: list[str], arg_name: str) -> tuple[str, bool]:
-    """Find a type entry in the input docstring that matches the given arg name."""
-    type_annotation = f":type {arg_name}: "
-    for line in lines:
-        if line.startswith(type_annotation):
-            return line, True
-    return type_annotation, False
-
-
-def _append_default(
-    app: Sphinx, lines: list[str], insert_index: int, type_annotation: str, formatted_default: str
-) -> str:
-    if app.config.typehints_defaults.endswith("after"):
-        # advance the index to the end of the :param: paragraphs
-        # (terminated by a line with no indentation)
-        # append default to the last nonempty line
-        nlines = len(lines)
-        next_index = insert_index + 1
-        append_index = insert_index  # last nonempty line
-        while next_index < nlines and (not lines[next_index] or lines[next_index].startswith(" ")):
-            if lines[next_index]:
-                append_index = next_index
-            next_index += 1
-        lines[append_index] += formatted_default
-
-    else:  # add to last param doc line
-        type_annotation += formatted_default
-
-    return type_annotation
-
-
-@dataclass
-class InsertIndexInfo:
-    insert_index: int
-    found_param: bool = False
-    found_return: bool = False
-    found_directive: bool = False
-
-
-# Sphinx allows so many synonyms...
-# See sphinx.domains.python.PyObject
-PARAM_SYNONYMS = ("param ", "parameter ", "arg ", "argument ", "keyword ", "kwarg ", "kwparam ")
-
-
-def get_insert_index(app: Sphinx, lines: list[str]) -> InsertIndexInfo | None:
-    # 1. If there is an existing :rtype: anywhere, don't insert anything.
-    if any(line.startswith(":rtype:") for line in lines):
-        return None
-
-    # 2. If there is a :returns: anywhere, either modify that line or insert just before it.
-    for at, line in enumerate(lines):
-        if line.startswith((":return:", ":returns:")):
-            return InsertIndexInfo(insert_index=at, found_return=True)
-
-    # 3. Insert after the parameters.
-    settings = get_default_settings(RSTParser)  # type: ignore[arg-type]
-    settings.env = app.env
-    doc = _safe_parse("\n".join(lines), settings)
-
-    for child in doc.children:
-        if _tag_name(child) != "field_list":
-            continue
-
-        if not any(c.children[0].astext().startswith(PARAM_SYNONYMS) for c in child.children):
-            continue
-
-        # Found it! Try to insert before the next sibling. If there is no next
-        # sibling, insert at end.
-        # If there is a next sibling but we can't locate a line number, insert
-        # at end. (I don't know of any input where this happens.)
-        next_sibling = child.next_node(descend=False, siblings=True)
-        line_no = _node_line_no(next_sibling) if next_sibling else None
-        at = max(line_no - 2, 0) if line_no else len(lines)
-        return InsertIndexInfo(insert_index=at, found_param=True)
-
-    # 4. Insert before examples
-    for child in doc.children:
-        if _tag_name(child) in {"literal_block", "paragraph", "field_list"}:
-            continue
-        line_no = _node_line_no(child)
-        at = max(line_no - 2, 0) if line_no else len(lines)
-        if lines[at - 1]:
-            break
-        return InsertIndexInfo(insert_index=at, found_directive=True)
-
-    # 5. Otherwise, insert at end
-    return InsertIndexInfo(insert_index=len(lines))
-
-
-def _safe_parse(inputstr: str, settings: Values | optparse.Values) -> nodes.document:
-    """
-    Parse RST without triggering extension directive side-effects.
-
-    Replaces non-builtin directive lookups with a no-op handler during parsing
-    to prevent duplicate ID registration and other side-effects from third-party
-    extensions like sphinx-needs.
-    """
-    original_lookup = directives.directive
-
-    def _safe_directive_lookup(
-        directive_name: str,
-        language_module: Any,
-        document: Any,
-    ) -> tuple[type[Directive] | None, list[Any]]:
-        cls, messages = original_lookup(directive_name, language_module, document)
-        if cls is not None and directive_name not in _BUILTIN_DIRECTIVES:
-            return _NoOpDirective, messages
-        return cls, messages
-
-    doc = new_document("", settings=settings)  # ty: ignore[invalid-argument-type]
-    with sphinx_domains(settings.env):
-        directives.directive = _safe_directive_lookup  # type: ignore[assignment]
-        try:
-            parser = _RstSnippetParser()
-            parser.parse(inputstr, doc)
-        finally:
-            directives.directive = original_lookup
-    return doc
-
-
-class _NoOpDirective(Directive):
-    has_content = True
-    optional_arguments = 99
-    final_argument_whitespace = True
-
-    def run(self) -> list[nodes.Node]:  # noqa: PLR6301
-        return []
-
-
-def _node_line_no(node: Node) -> int | None:
-    if node is None:
-        return None
-    while node.line is None and node.children:
-        node = node.children[0]
-    return node.line
-
-
-def _tag_name(node: Node) -> str:
-    return node.tagname
-
-
-def _inject_rtype(  # noqa: C901, PLR0911, PLR0913, PLR0917
+def _inject_rtype(  # noqa: PLR0913, PLR0917
     type_hints: dict[str, Any],
     original_obj: Any,
     app: Sphinx,
     what: str,
     name: str,
     lines: list[str],
+    fmt: Any,
 ) -> None:
     if inspect.isclass(original_obj) or inspect.isdatadescriptor(original_obj):
         return
-    if what == "method" and name.endswith(".__init__"):  # avoid adding a return type for data class __init__
+    if what == "method" and name.endswith(".__init__"):
         return
     if not app.config.typehints_document_rtype:
         return
@@ -1211,13 +337,8 @@ def _inject_rtype(  # noqa: C901, PLR0911, PLR0913, PLR0917
     if _has_yields_section(lines) and _is_generator_type(type_hints["return"]):
         return
 
-    r = get_insert_index(app, lines)
+    r = fmt.get_rtype_insert_info(app, lines)
     if r is None:
-        return
-
-    insert_index = r.insert_index
-
-    if not app.config.typehints_use_rtype and r.found_return and " -- " in lines[insert_index]:
         return
 
     short_literals = app.config.python_display_short_literal_types
@@ -1225,39 +346,7 @@ def _inject_rtype(  # noqa: C901, PLR0911, PLR0913, PLR0917
         format_annotation(type_hints["return"], app.config, short_literals=short_literals)
     )
 
-    if r.found_param and insert_index < len(lines) and lines[insert_index].strip():
-        insert_index -= 1
-
-    if insert_index > 0 and insert_index <= len(lines) and lines[insert_index - 1].strip():
-        # ensure that :rtype: doesn't get joined with a paragraph of text
-        lines.insert(insert_index, "")
-        insert_index += 1
-
-    if app.config.typehints_use_rtype or not r.found_return:
-        line = f":rtype: {formatted_annotation}"
-        lines.insert(insert_index, line)
-        if r.found_directive:
-            lines.insert(insert_index + 1, "")
-    else:
-        line = lines[insert_index]
-        lines[insert_index] = f":return: {formatted_annotation} --{line[line.find(' ') :]}"
-
-
-_GENERATOR_TYPES = frozenset({
-    collections.abc.Generator,
-    collections.abc.Iterator,
-    collections.abc.AsyncGenerator,
-    collections.abc.AsyncIterator,
-})
-
-
-def _is_generator_type(annotation: Any) -> bool:
-    origin = getattr(annotation, "__origin__", None)
-    return origin in _GENERATOR_TYPES or annotation in _GENERATOR_TYPES
-
-
-def _has_yields_section(lines: list[str]) -> bool:
-    return any(line.lstrip().startswith((":Yields:", ":yields:", ":yield:")) for line in lines)
+    fmt.inject_rtype(lines, formatted_annotation, r, use_rtype=app.config.typehints_use_rtype)
 
 
 def validate_config(app: Sphinx, env: BuildEnvironment, docnames: list[str]) -> None:  # noqa: ARG001
@@ -1272,20 +361,6 @@ def validate_config(app: Sphinx, env: BuildEnvironment, docnames: list[str]) -> 
         raise ValueError(msg)
 
 
-def unescape(escaped: str) -> str:
-    # For some reason the string we get has a bunch of null bytes in it??
-    # Remove them...
-    escaped = escaped.replace("\x00", "")
-    # For some reason the extra slash before spaces gets lost between the .rst
-    # source and when this directive is called. So don't replace "\<space>" =>
-    # "<space>"
-    return re.sub(r"\\([^ ])", r"\1", escaped)
-
-
-def add_type_css_class(type_rst: str) -> str:
-    return f":sphinx_autodoc_typehints_type:`{rst.escape(type_rst)}`"
-
-
 def sphinx_autodoc_typehints_type_role(
     _role: str,
     _rawtext: str,
@@ -1295,12 +370,6 @@ def sphinx_autodoc_typehints_type_role(
     _options: dict[str, Any] | None = None,
     _content: list[str] | None = None,
 ) -> tuple[list[Node], list[Node]]:
-    """
-    Add css tag around rendered type.
-
-    The body should be escaped rst. This renders its body as rst and wraps the
-    result in <span class="sphinx_autodoc_typehints-type"> </span>
-    """
     unescaped = unescape(text)
     doc = parse(unescaped, inliner.document.settings)
     n = nodes.inline(text)
@@ -1323,7 +392,7 @@ def setup(app: Sphinx) -> dict[str, bool]:
     app.add_config_value("typehints_use_signature_return", False, "env")  # noqa: FBT003
     app.add_config_value("typehints_fixup_module_name", None, "env")
     app.add_role("sphinx_autodoc_typehints_type", sphinx_autodoc_typehints_type_role)
-    app.connect("env-before-read-docs", validate_config)  # config may be changed after “config-inited” event
+    app.connect("env-before-read-docs", validate_config)
     app.connect("autodoc-process-signature", process_signature)
     app.connect("autodoc-process-docstring", process_docstring)
     install_patches(app)
