+++ tests/roots/test-issue_347/conf.py
@@ -1,0 +1,16 @@
+from __future__ import annotations
+
+import pathlib
+import sys
+
+sys.path.insert(0, str(pathlib.Path(__file__).parent))
+
+master_doc = "index"
+
+extensions = [
+    "sphinx.ext.autodoc",
+    "numpydoc",
+    "sphinx_autodoc_typehints",
+]
+# Disable numpydoc's class members listing to avoid interfering with autodoc
+numpydoc_show_class_members = False
