+++ tests/roots/test-issue_347/index.rst
@@ -1,0 +1,2 @@
+Test
+====
