+++ tests/test_integration_issue_572.py
@@ -7,7 +7,10 @@ from unittest.mock import patch
 
 import pytest
 
-from sphinx_autodoc_typehints import _get_type_hint
+from sphinx_autodoc_typehints._resolver import _get_type_hint
+
+if sys.version_info >= (3, 14):
+    import annotationlib
 
 if TYPE_CHECKING:
     from io import StringIO
@@ -40,14 +43,12 @@ def test_forward_ref_builds_without_errors(
 @needs_314
 def test_get_type_hint_uses_annotationlib_on_name_error() -> None:
     """_get_type_hint falls back to annotationlib.get_annotations on 3.14+ NameError."""
-    import annotationlib  # noqa: PLC0415
-
     sentinel = {"x": int}
 
     def dummy() -> None: ...
 
     with (
-        patch("sphinx_autodoc_typehints.get_type_hints", side_effect=NameError("name 'Foo' is not defined")),
+        patch("sphinx_autodoc_typehints._resolver.get_type_hints", side_effect=NameError("name 'Foo' is not defined")),
         patch.object(annotationlib, "get_annotations", return_value=sentinel) as mock_get_ann,
     ):
         result = _get_type_hint([], "dummy", dummy, {})
@@ -57,12 +58,12 @@ def test_get_type_hint_uses_annotationlib_on_name_error() -> None:
 
 
 @pytest.mark.skipif(sys.version_info >= (3, 14), reason="Tests pre-3.14 fallback path")
-def test_get_type_hint_falls_back_to_dunder_annotations_before_314() -> None:
+def test_get_type_hint_falls_back_to_dunder_annotations_before_314() -> None:  # pragma: <3.14 cover
     """_get_type_hint falls back to __annotations__ on pre-3.14 NameError."""
 
     def dummy(x: int) -> str: ...
 
-    with patch("sphinx_autodoc_typehints.get_type_hints", side_effect=NameError("name 'Foo' is not defined")):
+    with patch("sphinx_autodoc_typehints._resolver.get_type_hints", side_effect=NameError("name 'Foo' is not defined")):
         result = _get_type_hint([], "dummy", dummy, {})
 
     assert result == dummy.__annotations__
