+++ tests/test_version.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
+from sphinx_autodoc_typehints import __version__
+
 
 def test_version() -> None:
-    from sphinx_autodoc_typehints import __version__  # noqa: PLC0415
-
     assert __version__
